modified the bash scripting and tmux articles
The redirection table in the bash scripting article was fixed and some
relevant links were added. Using regexes along with the BASH_REMATCH
variable was also specified. The changes in this article were staged
using `git add --patch`.

The tmux article was modified to add a method to rearrange tabs.